[PIR+CINN]Refactor BuildCinnPass Logic
