+++ paddle/cinn/hlir/framework/pir/utils.cc
@@ -14,18 +14,27 @@
 
 #include "paddle/cinn/hlir/framework/pir/utils.h"
 
+#include <regex>
 #include <string>
 #include <unordered_map>
 #include "glog/logging.h"
 
+#include "paddle/cinn/hlir/dialect/operator/ir/manual_op.h"
 #include "paddle/cinn/hlir/dialect/operator/ir/op_dialect.h"
 #include "paddle/cinn/hlir/framework/op.h"
 #include "paddle/cinn/hlir/framework/pir/op_mapper.h"
 #include "paddle/fluid/pir/dialect/operator/ir/op_attribute.h"
+#include "paddle/fluid/pir/dialect/operator/ir/pd_op.h"
 #include "paddle/phi/common/data_type.h"
 #include "paddle/pir/core/builtin_op.h"
 #include "paddle/pir/core/builtin_type.h"
+#include "paddle/pir/dialect/control_flow/ir/cf_dialect.h"
+#include "paddle/pir/dialect/control_flow/ir/cf_op.h"
 #include "paddle/pir/dialect/shape/ir/shape_attribute.h"
+#include "paddle/utils/flags.h"
+
+PD_DECLARE_string(allow_cinn_ops);
+PD_DECLARE_string(deny_cinn_ops);
 
 namespace cinn {
 namespace hlir {
@@ -45,16 +54,161 @@ const std::unordered_map<std::string, std::string> CompatibleInfo::OP_NAMES = {
     {"pd_op.maximum", "max"},
     {"pd_op.minimum", "min"},
     {"pd_op.split_with_num", "split"},
-    {"pd_op.reshape", "reshape"},
     {"cinn_op.reshape", "reshape"},
     {"cinn_op.scale", "scale"},
     {"cinn_op.broadcast", "broadcast_to"},
-    {"pd_op.expand", "broadcast_to"},
-    {"cinn_op.generate_shape", "generate_shape"},
     // The following should implement OpPattern in pd_to_cinn_pass,
     // otherwise, it will be block in BuildCinnPass.
     {"cinn_op.squeeze", ""},
     {"cinn_op.unsqueeze", ""}};
+
+namespace {
+using GroupOpsVec = std::vector<::pir::Operation*>;
+// The delim(`;`) that is used to split the FLAGS_allow_cinn_ops
+// & FLAGS_deny_cinn_ops.
+constexpr char kDelim[] = ";";
+
+// OpTransInfo contains informations used to detect subgraphs
+// supported by the CINN compiler.
+class OpTransInfo {
+  using DeParamCondT =
+      std::unordered_map<std::string, std::unordered_set<std::string>>;
+
+ public:
+  OpTransInfo() {}
+
+  const DeParamCondT& deny_param_cond() const { return deny_param_cond_; }
+  const std::unordered_set<std::string>& default_deny_ops() const {
+    return default_deny_ops_;
+  }
+
+ private:
+  DeParamCondT deny_param_cond_{{"batch_norm", {"ReserveSpace"}},
+                                {"batch_norm_grad", {"ReserveSpace"}}};
+
+  std::unordered_set<std::string> default_deny_ops_{
+      "feed", "fetch", "conv2d", "conv2d_grad", "dropout"};
+};
+
+std::unordered_set<std::string> StringSplit(const std::string& str,
+                                            const std::string& delim) {
+  std::regex reg(delim);
+  std::unordered_set<std::string> elems{
+      std::sregex_token_iterator(str.begin(), str.end(), reg, -1),
+      std::sregex_token_iterator()};
+  elems.erase("");
+  return elems;
+}
+
+std::string GetDebugInfo(const std::unordered_set<std::string>& names) {
+  std::string debug_info = "[";
+  for (auto& name : names) {
+    debug_info.append(name);
+    debug_info.append(", ");
+  }
+  debug_info.append("]");
+  return debug_info;
+}
+
+bool IsSupportForCinn(const ::pir::Operation& op);
+
+// In case of op has some attributes generated by FullOp, it need
+// implement OpPattern in pd_to_cinn_pass. Otherwise, we mark them
+// as unimplement ops.
+bool UnimplementOps(const ::pir::Operation& op) {
+  // cinn not support uniform, the FullOp of max and min support NOT generate by
+  // CINN
+  if (op.isa<paddle::dialect::FullOp>()) {
+    auto out = op.result(0);
+    if (out.use_count() > 0) {
+      return !IsSupportForCinn(*(out.first_use().owner()));
+    }
+  }
+  return false;
+}
+
+bool HaveZeroDimInput(const ::pir::Operation& op) {
+  bool have_zero_dim = false;
+  for (size_t i = 0; i < op.num_operands(); ++i) {
+    auto in = op.operand_source(i);
+    if (in) {
+      if (auto tensor_type =
+              in.type().dyn_cast<paddle::dialect::DenseTensorType>()) {
+        if (tensor_type.dims().size() == 0) {
+          have_zero_dim = true;
+        }
+      }
+    }
+  }
+
+  return have_zero_dim;
+}
+
+bool AllInputDenseTensor(const ::pir::Operation& op) {
+  bool all_denese_tensor = true;
+  for (size_t i = 0; i < op.num_operands(); ++i) {
+    auto in = op.operand_source(i);
+    if (in) {
+      if (!(in.type().isa<paddle::dialect::DenseTensorType>())) {
+        all_denese_tensor = false;
+      }
+    }
+  }
+
+  return all_denese_tensor;
+}
+
+bool IsRegisteredInCINN(const ::pir::Operation& op) {
+  if (CompatibleInfo::OP_NAMES.find(op.name()) !=
+      CompatibleInfo::OP_NAMES.end()) {
+    return true;
+  }
+  // After PdToCinnPass, if pd_op.reshape still exists, return false.
+  std::string black_op_name =
+      std::string(cinn::dialect::OperatorDialect::name()) + "." +
+      CompatibleInfo::OpName(op);
+  if (CompatibleInfo::OP_NAMES.find(black_op_name) !=
+      CompatibleInfo::OP_NAMES.end()) {
+    VLOG(4) << "Found black op after PdToCinnPass, because it has Attribute "
+               "Tensor: "
+            << op.name();
+    return false;
+  }
+  return OpRegistry::Global()->Find(CompatibleInfo::OpName(op)) != nullptr;
+}
+
+bool IsSupportForCinn(const ::pir::Operation& op) {
+  if (!AllInputDenseTensor(op) || HaveZeroDimInput(op) || UnimplementOps(op)) {
+    return false;
+  }
+  auto allow_ops = StringSplit(FLAGS_allow_cinn_ops, kDelim);
+  auto deny_ops = StringSplit(FLAGS_deny_cinn_ops, kDelim);
+  VLOG(4) << "The allowed Cinn Ops: " << GetDebugInfo(allow_ops);
+  VLOG(4) << "The denied Cinn Ops: " << GetDebugInfo(deny_ops);
+  // Strip the dialect, like pd_op.abs -> abs
+  const auto op_name = CompatibleInfo::OpName(op);
+
+  OpTransInfo trans_info;
+  bool is_support =
+      IsRegisteredInCINN(op) && !trans_info.default_deny_ops().count(op_name);
+  VLOG(4) << op_name << " is_support: " << is_support << " "
+          << IsRegisteredInCINN(op);
+  // if the op type is registered in CINN and allow_ops is not empty, return
+  // true only when it is in allow_ops
+  if (!allow_ops.empty()) {
+    return is_support && allow_ops.count(op_name);
+  }
+  // if the op type is registered in CINN and deny_ops is not empty, return
+  // true only when it is not in deny_ops
+  if (!deny_ops.empty()) {
+    return is_support && !deny_ops.count(op_name);
+  }
+
+  // if the user doesn't set FLAGS_allow_cinn_ops and FLAGS_deny_cinn_ops,
+  // return true only when it is registered in CINN
+  return is_support;
+}
+}  // namespace
 
 // In following cases, the op is marked SupportCinn:
 // 1. its name is in OP_NAMES, like pd_op.sum;
@@ -62,19 +216,7 @@ const std::unordered_map<std::string, std::string> CompatibleInfo::OP_NAMES = {
 //    Such as cinn_op.reshape, except pd_op.reshape;
 // 3. otherwise, it should be registered in OpRegistry;
 bool CompatibleInfo::IsSupportCinn(const ::pir::Operation& op) {
-  if (OP_NAMES.find(op.name()) != OP_NAMES.end()) {
-    return true;
-  }
-  // After PdToCinnPass, if pd_op.reshape still exists, return false.
-  std::string black_op_name =
-      std::string(cinn::dialect::OperatorDialect::name()) + "." + OpName(op);
-  if (OP_NAMES.find(black_op_name) != OP_NAMES.end()) {
-    VLOG(4) << "Found black op after PdToCinnPass, because it has Attribute "
-               "Tensor: "
-            << op.name();
-    return false;
-  }
-  return OpRegistry::Global()->Find(OpName(op)) != nullptr;
+  return IsSupportForCinn(op);
 }
 
 std::string CompatibleInfo::OpName(const ::pir::Operation& op) {
